+++ src/Image/Image.php
@@ -3,18 +3,24 @@
  * @copyright Copyright (c) Wizacha
  * @license Proprietary
  */
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace Wizaplace\SDK\Image;
 
+use GuzzleHttp\Psr7\Uri;
+use Psr\Http\Message\UriInterface;
+use Wizaplace\SDK\Exception\NotFound;
+
 /**
- * Class Image
- * @package Wizaplace\SDK\Image
+ * Class Image.
  */
 final class Image
 {
     /** @var int */
     private $id;
+
+    /** @var UriInterface[] */
+    private $urls;
 
     /**
      * @internal
@@ -24,6 +30,10 @@ final class Image
     public function __construct(array $data)
     {
         $this->id = $data['id'];
+
+        foreach ($data['urls'] as $key => $url) {
+            $this->urls[$key] = new Uri($url);
+        }
     }
 
     /**
@@ -32,5 +42,40 @@ final class Image
     public function getId(): int
     {
         return $this->id;
+    }
+
+    /**
+     * @return UriInterface[]
+     */
+    public function getUrls(): array
+    {
+        return $this->urls;
+    }
+
+    public function getUrl(string $size = 'original'): UriInterface
+    {
+        if (false === \array_key_exists($size, $this->urls)) {
+            throw new NotFound(
+                "Invalid size '$size'. Available sizes: ".implode(', ', array_keys($this->urls)).'.'
+            );
+        }
+
+        return $this->urls[$size];
+    }
+
+    /**
+     * @return array
+     */
+    public function jsonSerialize(): array
+    {
+        return [
+            'id' => $this->id,
+            'urls' => array_map(
+                function ($uri) {
+                    return (string) $uri;
+                },
+                $this->urls
+            ),
+        ];
     }
 }